change string lengths from type int to size_t

+++ chash.h
@@ -2,6 +2,6 @@
 
 struct chash_t;
 
-struct chash_t *chash_create(char **keys, int nkeys, int replicas);
-char *chash_lookup(struct chash_t *chash, char *key, int len);
+struct chash_t *chash_create(char **keys, size_t nkeys, int replicas);
+char *chash_lookup(struct chash_t *chash, char *key, size_t len);
 void chash_free(struct chash_t *chash);
